update: Full workflow & folder organization
