+++ pyproject.toml
@@ -6,7 +6,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "a2a-sdk>=1.1.0",
-    "agno==2.5.14",
+    "agno==2.6.6",
     "bs4>=0.0.2",
     "chromadb>=1.5.8",
     "duckdb==1.3.0",
